Split tests for UnloadVirtualEndPointAndItemData into collection and object tests. (RM-7236)
git-svn-id: https://svn.re-motion.org/svn/Remotion/trunk@32606 cb0deee2-328e-b949-94ea-fb3139fab790
